Finish Password and Confirm password validation
Created a Javascript validation for both password and confirm password.
It toggle the border style of both password and confirm password input
to red and also attach a class unmatched to the span to show a visual
alert to inform the user that the password do not match.
It also toggle the type of button from submit to a generic button